Provided test for detect_buffers and multi
I added a test to check multi command when detect_buffers is enabled.
I had also written a test to check that multi return buffers when
detect_buffers is enabled, but it doesn't so it seems another issue to
open, so I left it commented for test passes and assume that
detect_buffers doesn't apply when multi command is used.