Fixed test - Wrong redis key and buffer check
I fixed the test that I wrote it wrong, so I was requesting a key that
it isn't exist and comparing a Stringify buffer with a wrong value.

I've realised that the test passed before, in my machine, because that
key existed in my redis server, so connect to a testing redis server and
emit ```flushall``` and after close the client connection (one test
check the number of clients connected) shouldn't give a false positive.