Remove bad .eval implementation
The implementation is not really good as mentioned in #722 and we pipline our commands.
That way we can't just replace the eval function as it was. This could result in violating the order of execution!

If we want to include a function like this we should not break the order of execution and also not recalculate the sha1 hash each time.