Adds config variable `enable_offline_queue_until_first_ready`
Configuration variable added that allows you to enable the offline
queue only while the connection is being opened, before it first
becomes ready. After it is ready, if you have `enable_offline_queue`
set to false then it will stop queuing. Useful for when you don’t want
an offline queue during normal operation (ie, you are using redis as a
cache that has a fallback to a primary data source) but you do want it
to queue requests initially when loading the connection without having
to listen to the ‘ready’ event being emitted.