chore(release): use deploy keys for relese (#3013)
main branch is protected and does not allow direct pushes.
the release action needs to push.
branch protection rules can be bypassed for people and apps, but not
github actions.

one of the workarounds is to use a ruleset in which we set a deploy key

see: https://github.com/orgs/community/discussions/25305\#discussioncomment-10728028