chore(build): fix npm run build command
importing from 'package.json' messes typescript,
as it starts thinking it depends on the files it wants to
write, resulting in build errors if the dist folders are not deleted.
ultimately it was very annoying to delete all dist folders and then build
when you just want to build. and its also much slower because
you loose all your cache